added more explanation

+++ Project_2/lr_main.py
@@ -69,7 +69,7 @@ def get_aprox_prediction(weights): # Returns the P(Y|W,X) from in the PDF
     return predict
 
 def get_likelihood(Yl, weights, Xljs):
-    return int(Yl)* np.dot(weights,Xljs[0])- np.log(1 + np.exp(np.dot(weights,Xljs[0])))# Conditional likelihood equation from PDF
+    return int(Yl)* np.dot(weights,Xljs[0])- np.log(1 + np.exp(np.float128(np.dot(weights,Xljs[0]))))# Conditional likelihood equation from PDF
 
 
 def minimize_likelihood(last_likelihood, weights_matrix):
@@ -157,3 +157,4 @@ weights = build_weights()
 print(get_aprox_prediction(weights)) # This pirnts out the final predictions where rows are the classes (news goups) and  columns are 
 # the training examples (documents used in training), so each value shown is the probability that the document (column) is from news group (row)
 # for the small set it prints as expected for the final weights however there is an overflow issue for the large database
+# so we might have to normalize the large data set in some way in order to bring down the size of the values
